third_party/blink/web_tests/external/wpt/css/css-tables/fixed-layout-2-expected.txt

This is a testharness.js-based test.
[FAIL] max-content
  assert_equals: Table is NOT in fixed mode expected 100 but got 50
Harness: the test ran to completion.
